backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
apscheduler==3.10.4
alembic==1.14.0

# OCR - OCR.space API (Free, no billing required)
Pillow==11.0.0
python-dateutil==2.9.0
requests==2.32.3

# Notifications
sendgrid==6.11.0  # Email (optional)
twilio==9.0.0     # SMS (optional)

# Google OAuth
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# AI - Gemini for intelligent date extraction
google-generativeai==0.8.3

# Payment Gateway - Razorpay
razorpay==1.4.1

# For Python 3.13 compatibility
setuptools>=75.0.0
